+++ src/app/locales/index.ts
@@ -1,5 +1,6 @@
 export * from "./types";
 export * from "./pr-body.builder";
 export * from "./locale-guard-labels.util";
-export * from "./pt-br.locale";
-export * from "./ru.locale";
+export * from "./repairs.registry";
+export { ptBrLocale } from "./pt-br/locale";
+export { ruLocale } from "./ru/locale";